Nothing wrong with an int here, so make it explicit and kill the warning.
git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/trunk@91217 13f79535-47bb-0310-9956-ffa450edef68

+++ server/util_time.c
@@ -157,7 +157,7 @@ static apr_status_t cached_explode(apr_exploded_time_t *xt, apr_time_t t,
         memcpy(&(cache_element->xt), xt, sizeof(apr_exploded_time_t));
         cache_element->t_validate = seconds;
     }
-    xt->tm_usec = t % APR_USEC_PER_SEC;
+    xt->tm_usec = (int)(t % APR_USEC_PER_SEC);
     return APR_SUCCESS;
 }
